# fix: use table instead of MV for section_video_engagement MV