### Description

This PR reduces the dependencies from the section/subsection_video_engagement MV, so that it doesn't get OOM killed on large datasets.

### Performance gains:
Before:
```
0 rows in set. Elapsed: 0.283 sec. Processed 11.21 thousand rows, 2.36 MB (39.64 thousand rows/s., 8.35 MB/s.)
Peak memory usage: 35.93 MiB
```
After:
```
0 rows in set. Elapsed: 0.062 sec. Processed 8.48 thousand rows, 1.97 MB (136.64 thousand rows/s., 31.79 MB/s.)
Peak memory usage: 12.89 MiB.
```