@openedx-webhooks 2024-06-14
@Ian2012 Even though your pull request wasn’t merged, please take a moment  to [answer a two question survey](https://docs.google.com/forms/d/e/1FAIpQLSceJOyGJ6JOzfy6lyR3T7EW_71OWUnNQXp68Fymsk3MkNoSDg/viewform?usp=pp_url&entry.1671973413=openedx/aspects-dbt&entry.867055334=https://github.com/openedx/aspects-dbt/pull/99&entry.1484655318=https://github.com/Ian2012&entry.752974735=2024-05-28+17:36&entry.1917517419=2024-06-14+21:45&entry.2133058324=No) so we can improve your experience in the future.

<!-- comment:end_survey -->